datastuff/criteo/requirements.txt
boto3
click
pyarrow
tqdm
avilabsutils-cprint
types-requests
